chore(dx): ツールチェーン整備 — dev 依存一本化・ruff B/RUF・seaborn 削除・Any-gate CI・CJK 回帰テスト (#1790)
* chore(deps): dev 依存を dependency-groups へ一本化し seaborn を削除 (plan 024)

pytest / ruff が [project.optional-dependencies].dev と
[dependency-groups].dev に二重宣言されドリフトしていたため
[dependency-groups].dev へ一本化した。素の \`uv sync\` で ruff が
入るようになったため --extra dev 指定を CI / README / ONBOARDING /
flake.nix から除去した。あわせて import ゼロで宣言のみ残っていた
直接依存 seaborn を削除した（pandas / matplotlib は独立宣言済みで
transitive にも不要）。


* chore(lint): ruff B/RUF 導入で検出したバックログを機械的に解消 (plan 024)

ruff select に B（flake8-bugbear）/ RUF を追加した際に検出された
約190件のバックログを解消した。RUF100（unused-noqa）/ RUF022（dunder-all
ソート）等は --fix で自動修正、B904（except 内 raise から from 欠落）/
RUF013（implicit Optional）/ RUF012（mutable class default → ClassVar）/
RUF005（list 結合 → unpacking）/ B905（zip strict 未指定）/ B007（未使用
ループ変数）等は個別に手動修正した。B905 は各 zip 呼び出しについて
呼び出し元のデータ構造から同長が自明なものだけ strict=True にし、
そうでないもの（呼び出し側が意図的に非同長を渡すケースがある等）は
strict=False（zip 既定と同じ挙動）にした。


* chore(ci): 新規 Any 使用ゲートに CI バックストップを追加 (plan 024)

.lefthook/pre-push/any-usage-gate.sh は origin/main 不在時に self-skip
していたため、client-side lefthook のみでは実質 advisory だった。CI に
any-gate job（pull_request イベントのみ、base sha を PRE_PUSH_DIFF_BASE
として既存スクリプトへ渡す）を追加し、強制力のあるゲートにした。
lefthook.yml の pre-push は変更しない（既存の 1 コマンド構成を維持）。


* test(deps): japanize-matplotlib の font 登録に対する glyph 回帰テストを追加 (plan 024)

japanize-matplotlib は2020年から未更新で、matplotlib のメジャー更新
（現行 3.10.8）で font 登録が壊れるリスクがある。壊れた場合の症状
（日本語ラベルの豆腐化）を matplotlib が描画時に出す
"Glyph NNNN ... missing from font(s)" UserWarning で検知する回帰テスト
tests/test_japanize_font_regression.py を追加した。現状のコードでは
pass することを確認済み（既存の font 登録は壊れていない）。


* docs(plans): plan 024 のステータスを DONE に更新


---------

+++ src/youtube_automation/scripts/fix_per_theme_timestamps.py
@@ -107,7 +107,7 @@ def compute_pattern_starts(music_dir: Path, patterns: dict[str, str]) -> list[tu
         letter = m.group(1).lower()
         # start time of this track in master = cumulative
         if letter not in seen:
-            seen[letter] = int(round(cumulative))
+            seen[letter] = round(cumulative)
         dur = get_duration(f)
         cumulative += dur
         if idx < len(files) - 1:
